hide ffmpeg ncvisual definition
